seventh commit

+++ StorageManage/StorageManage/ButtonClick/DeleteMalfunction.cs
@@ -24,11 +24,13 @@ namespace StorageManage.ButtonClick
             if (DRV == null) { MessageBox.Show("Удаление прервано, Вы не выбрали запись для удаления."); return; }
             DataRow DR = DRV.Row;
             object[] arr = DR.ItemArray;
-            //MySqlDataReader reader = window.ex.returnResult("select idorders from orders where iddetails=" + arr[0]);
-            //if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
-            //window.ex.closeCon();
-            
+            MySqlDataReader reader = window.ex.returnResult("select recordid from repairorders_malfunctions where idmalfunctions=" + arr[0]);
+            if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
+            window.ex.closeCon();
             window.ex.ExecuteWithoutRedaer("delete from malfunctions where idmalfunctions=" + arr[0]);
+            window.ex.ExecuteWithoutRedaer("delete from malfunctions_causes where idmalfunctions=" + arr[0]);
+            window.ex.ExecuteWithoutRedaer("delete from malfunctions_details where idmalfunctions=" + arr[0]);
+
 
             DataGridUpdater.MalfunctionsDataGridUpdate(window);
         }
